client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"streamlit>=1.0",
	"requests>=2.0",
	"pillow>=9.0",
]
